Calculated a Tie/Added Messaging for Player
When Player X nor Player O win the game, I put a condition in my code to
relay this to the user, by stating that the game was a tie. Also made it
so that when a user clicks on an occupied space a message appears,
letting the user know to choose a different space.